pyproject.toml
[project]
name = "a430sysid"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Xudong Gong", email = "gongxudong_cs@aliyun.com" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a430py",
    "gymnasium>=1.2.0",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "optuna>=4.5.0",
    "scipy>=1.16.2",
]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.uv.sources]
a430py = { path = "../a430py", editable = true }

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pandas>=2.3.2",
    "pre-commit>=4.3.0",
    "pysindy>=2.0.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "seaborn>=0.13.2",
]
